requirements.txt
livekit
livekit-api
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-noise-cancellation
livekit-plugins-turn-detector

asyncpg
psycopg2-binary
alembic

fastapi
uvicorn[standard]
websockets
python-multipart

python-dotenv
pydantic
pydantic-settings

openai

sentry-sdk[fastapi]
prometheus-client
redis

textblob
aiofiles
tenacity
httpx

pytest
pytest-asyncio
pytest-cov